move validate compose code to compose
